requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
django-lifecycle==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
urlman==2.0.1
